refactor: organize platform-specific code under src/platforms/
Move mobile UI components to src/platforms/mobile/:
- MobileShell, MobileAppLifecycle, MobileSidebarShell, use-is-mobile

Move desktop UI components to src/platforms/desktop/:
- ElectronDragRegion, DiscordPresenceSync, use-desktop-app, use-download-links

Add barrel exports for all three platforms (mobile, desktop, smart-tv).
No UI or behavior changes — file moves + import updates only.
