+++ src/features/search/components/content-detail-modal.tsx
@@ -6,8 +6,8 @@ import { useTranslations } from 'next-intl';
 import { useEffect, useRef } from 'react';
 import { Button } from '@/components/ui/button';
 import { PlaybackCountdown } from '@/features/watch/components/PlaybackCountdown';
-import { useIsMobile } from '@/hooks/use-is-mobile';
 import { cn, getOptimizedImageUrl } from '@/lib/utils';
+import { useIsMobile } from '@/platforms/mobile/use-is-mobile';
 import { useContentDetailModal } from '../hooks/use-content-detail-modal';
 import { ContentType } from '../types';
 import { ContentActions, ContentInfo } from './content-info';
