+++ src/platforms/desktop/DiscordPresenceSync.tsx
