+++ src/platforms/desktop/index.ts
@@ -1,0 +1,4 @@
+export { DiscordPresenceSync } from './DiscordPresenceSync';
+export { ElectronDragRegion } from './ElectronDragRegion';
+export { useDesktopApp } from './use-desktop-app';
+export { useDownloadLinks } from './use-download-links';
