+++ src/platforms/index.ts
@@ -1,0 +1,3 @@
+export * from './desktop';
+export * from './mobile';
+export * from './smart-tv';
