+++ src/platforms/mobile/MobileAppLifecycle.tsx
